Create ABSTRACT
In recent times, Internet of things (IoT) has been a significant source for technological solutions in many industries, that enable the combination of real-world data and services into the current information networking technologies. On the other hand, the deployment of Wireless Sensor Network technology with reference to Internet of things has been deployed in an exceedingly big selection of applications, this can be because sensors are getting smaller in size and the cost of production is comparatively cheap. This research paper is geared towards presenting a review on wireless sensor networks, highlighting the different types of sensor networks, their challenges, various applications, as well as Advantages/Disadvantages of WSN.